- buttons for make step - add fee for winner - multiChains support

+++ blockchain_app/contracts/IGame.sol
@@ -2,6 +2,6 @@
 pragma solidity ^0.8.9;
 
 interface IGame {
-    function init(uint _timeoutTime, address _token, uint _coins, uint _size) external;
+    function init(uint16 _timeoutTime, address _token, uint _coins, uint8 _size) external;
 
 }
